Change Urllib3 to work with test_base.py
